F1 Export Improvvvvements
Updated the SQL for Communications to get the most recent communication value per person per type.  It also joins back on the household table to set the indiviual_Id when it's only a household communication.

Fixed the SQL for getting Groups to include Groups that aren't included in the Group Description Table

Added the Default Tag Comment as an exported Attribute

Modified School Key to be F1 School.  This make a new attribugte in Rock vs using the Rock Attribute.

Moved the Barcode to a PersonSearch Key vs an Attribute

Added Logic for creating Transcations with the type "Recepit"

Fixed a bug for when "is_Public" column is null in setting if a group is public or not.

Trimed the length of all person names to 50 characters

Add logic so that if a person doesn't have an 'Email' communication type, to see if they have an Infellowship Login.  If so, use that as email.

Added exporting the IsUnlisted property of a phone number.

Added logic to set Mobile or Cell phone number types to SMS enabled if they're listed.